Crear token

+++ resolver.js
@@ -7,6 +7,8 @@ const { buscarUsuarioPorIdentificacion} = require('./service/usuario.service')
 const Project = require('./model/proyectoModel')
 const User = require('./model/usuarioModel')
 let aes256 = require('aes256');
+const { isLider } = require('./middleware/authjwt');
+const jwt = require('jsonwebtoken')
 
 const listUsuarios = [
     {
@@ -38,8 +40,10 @@ const resolvers = {
         //usuarios: () => listUsuarios,
         usuarios: async () => await User.find({}),
         usuario: (parent, args, context, info) => buscarUsuarioPorIdentificacion(args.identificacion),
-        
-        proyectos: async () => proyectos(),
+        //proyectos: async () => proyectos(),
+        proyectos: async (parent, args, context, info) => {
+            return proyectos()
+        },
         getProject: async (parent, args, context, info) => getProject( args.nombre ),
     },
     Mutation: {
@@ -53,8 +57,7 @@ const resolvers = {
                 .catch(err => console.log(err));
             //.catch(err => "fallo la creacion");
         },
-        
-        activeUser: (parent, args, context, info) => {
+         activeUser: (parent, args, context, info) => {
             return User.updateOne({ identificacion: args.identificacion }, { estado: "Activo" })
                 .then(u => "Usuario Activo")
                 .catch(err => "Fallo la activacion");
@@ -66,12 +69,18 @@ const resolvers = {
         },
 
         deleteUser: (parent, args, context, info) => {
-            return User.deleteOne({ identificacion: args.ident })
-                .then(u => "Usuario eliminado")
-                .catch(err => "Fallo la eliminacion");
+            if (isLider(context.rol)) {
+                return User.deleteOne({ identificacion: args.ident })
+                    .then(u => "Usuario eliminado")
+                    .catch(err => "Fallo la eliminacion");
+            }
         },
-        deleteProject: (parent, args, context, info) => deleteProject(args.nombreProyecto),          
-        
+        deleteProject: (parent, args, context, info) => {
+            //if (isAdmin(context.rol)) {
+            if (isLider(context.rol)) {
+                deleteProject(args.nombreProyecto)
+            }
+        },
         insertUserToProject: async (parent, args, context, info) => addUserProject(args.identificacion, args.nombreProyecto),
         
         createUser: (parent, args, context, info) => {
@@ -83,9 +92,32 @@ const resolvers = {
                 .then(u => "usuario creado")
                 .catch(err => console.log(err));
         },
-        createProject: (parent, args, context, info) => createProject(args.project),                      
-        
-        
+        createProject: (parent, args, context, info) => {
+            if (isLider(context.rol)) { 
+            createProject(args.project)                      
+               
+     }
+    },
+       /* autenticar: async (parent, args, context, info) => {
+            try {
+                const usuario = await User.findOne({ email: args.usuario })
+                if (!usuario) {
+                    return "Verique usuario y contrasena"
+                }
+
+                const claveDesencriptada = aes256.decrypt(key, usuario.clave)
+                if (args.clave != claveDesencriptada) {
+                    return "Verique usuario y contrasena"
+                }
+                const token = jwt.sign({
+                    rolesito: usuario.perfil
+                }, key, { expiresIn: 60 * 60 * 2 })
+
+                return token
+            } catch (error) {
+                console.log(error)
+            }
+        }*/
     }
-}
+}     
 module.exports = resolvers
